pyproject.toml
[project]
name = "llm-function-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
]
